Part 3 Merge v1

+++ app/src/main/java/st10036509/countify/user_interface/counter/CounterViewFragment.kt
@@ -5,12 +5,16 @@ import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
+import android.widget.Toast
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.ItemTouchHelper
@@ -19,15 +23,19 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.ListenerRegistration
 import st10036509.countify.R
 import st10036509.countify.adapter.CounterAdapter
 import st10036509.countify.model.CounterModel
 import st10036509.countify.model.UserManager
+import st10036509.countify.service.CounterDatabaseHelper
 import st10036509.countify.service.FirebaseAuthService
+import st10036509.countify.service.FirestoreService
 import st10036509.countify.service.NavigationService
 import st10036509.countify.service.Toaster
 import st10036509.countify.user_interface.account.SettingsFragment
 import java.util.Locale
+
 
 class CounterViewFragment : Fragment() {
 
@@ -39,6 +47,27 @@ class CounterViewFragment : Fragment() {
     private val firestore = FirebaseFirestore.getInstance()
     private var counterList: MutableList<CounterModel> = mutableListOf()
     private var currentUser: FirebaseUser? = null
+    private var registration: ListenerRegistration? = null
+
+    // method to fetch all counters on page start
+    override fun onStart() {
+        super.onStart()
+        fetchCountersFromLocalDatabase()
+        fetchCountersFromFirestore() // fetch counters when the fragment starts
+    }
+
+    // method to remove the listener when the page is closed
+    override fun onStop() {
+        super.onStop()
+        registration?.remove() // remove the listener to avoid leaks
+    }
+    override fun onResume() {
+        super.onResume()
+        val dbHelper = CounterDatabaseHelper(context ?: return)
+        if (isConnected(context ?: return)) {
+            syncUnsyncedCounters()
+        }
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -50,13 +79,24 @@ class CounterViewFragment : Fragment() {
         recyclerView.layoutManager = LinearLayoutManager(context)
         currentUser = FirebaseAuthService.getCurrentUser()
 
-        fetchCountersFromFirestore()
+        //fetchCountersFromFirestore()
         setAppLocale(if (UserManager.currentUser?.language == 1) "af" else "default", requireContext())
 
         // Set up ItemTouchHelper for swipe-to-delete functionality
         setupSwipeToDelete()
 
-        //Initializing toaster
+        //initialize the counter list
+        counterList = mutableListOf()
+
+        //initialize the adapter
+        counterAdapter = CounterAdapter(counterList, this)
+
+        //set the adapter to your RecyclerView
+        recyclerView.adapter = counterAdapter
+
+        fetchCountersFromLocalDatabase()
+
+        //initializing toaster
         toaster = Toaster(this)
 
         return view
@@ -72,6 +112,8 @@ class CounterViewFragment : Fragment() {
         context.createConfigurationContext(config)
         context.resources.updateConfiguration(config, context.resources.displayMetrics)
     }
+
+
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -93,30 +135,101 @@ class CounterViewFragment : Fragment() {
         val userID = currentUser?.uid
         firestore.collection("counters")
             .whereEqualTo("userId", userID)
-            .get()
-            .addOnSuccessListener { result ->
-                counterList.clear()
-                for (document in result) {
-                    val counter = CounterModel(
-                        counterId = document.id,  // Assign the document ID
-                        name = document.getString("name") ?: "",
-                        changeValue = document.getLong("incrementValue")?.toInt() ?: 0,
-                        count = document.getLong("count")?.toInt() ?: 1,
-                        createdTimestamp = document.getLong("createdTimestamp") ?: 0L,
-                        repetition = document.getString("repetition") ?: "",
-                        userId = document.getString("userId") ?: ""
-                    )
-                    counterList.add(counter)
+            .addSnapshotListener { snapshot, e ->
+                if (e != null) {
+                    toaster.showToast(getString(R.string.counter_pull_failed))
+                    Log.w(TAG, "Listen failed.", e)
+                    return@addSnapshotListener
                 }
-                // Setup the adapter with the fetched counters
-                counterAdapter = CounterAdapter(counterList, this)
-                recyclerView.adapter = counterAdapter
-                //.showToast(getString(R.string.counter_pull_successful))
-            }
-            .addOnFailureListener { exception ->
-                toaster.showToast(getString((R.string.counter_pull_failed)))
+
+                if (snapshot != null) {
+                    counterList.clear()
+                    for (document in snapshot.documents) {
+                        val counter = CounterModel(
+                            counterId = document.id,
+                            name = document.getString("name") ?: "",
+                            changeValue = document.getLong("changeValue")?.toInt() ?: 0,
+                            count = document.getLong("count")?.toInt() ?: 1,
+                            startValue =  document.getLong("startValue")?.toInt() ?: 1,
+                            createdTimestamp = document.getLong("createdTimestamp") ?: 0L,
+                            repetition = document.getString("repetition") ?: "",
+                            userId = document.getString("userId") ?: "",
+                            synced = (document.getBoolean("synced") ?: 0) as Boolean,
+                            )
+                        counterList.add(counter)
+                    }
+                    //notify the adapter that the data has changed
+                    counterAdapter.notifyDataSetChanged()
+                } else {
+                    Log.d(TAG, "Current data: null")
+                }
             }
     }
+
+
+    fun syncUnsyncedCounters() {
+        val dbHelper = CounterDatabaseHelper(context ?: return)
+        val unsyncedCounters = dbHelper.getUnsyncedCounters()
+
+        if (isConnected(context ?: return)) {
+            for (counter in unsyncedCounters) {
+                FirestoreService.addCounter(counter) { success, error ->
+                    if (success) {
+                        // Mark counter as synced in the local SQLite database
+                        counter.synced = true
+                        dbHelper.updateCounter(counter)
+
+                        // Remove the synced counter from the SQLite database
+                        dbHelper.deleteCounter(counter.counterId)
+
+                        Toast.makeText(
+                            context,
+                            "Synced counter: ${counter.name}",
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    } else {
+                        Log.e(
+                            "syncUnsyncedCounters",
+                            "Failed to sync counter: ${counter.name} - $error"
+                        )
+                    }
+                }
+            }
+        }
+    }
+
+
+    fun isConnected(context: Context): Boolean {
+        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            val network = connectivityManager.activeNetwork ?: return false
+            val activeNetwork = connectivityManager.getNetworkCapabilities(network) ?: return false
+            return when {
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> true
+                else -> false
+            }
+        } else {
+            @Suppress("DEPRECATION")
+            val networkInfo = connectivityManager.activeNetworkInfo
+            return networkInfo != null && networkInfo.isConnected
+        }
+    }
+
+    private fun fetchCountersFromLocalDatabase() {
+        val dbHelper = CounterDatabaseHelper(context ?: return) // Pass the context of your activity/fragment
+        val userId = currentUser?.uid ?: return
+
+        // Fetch counters from SQLite
+        val counterList = dbHelper.getCountersByUserId(userId)
+
+        // Update your adapter or UI with the retrieved counter list
+        counterAdapter.updateData(counterList)
+        counterAdapter.notifyDataSetChanged()
+    }
+
 
     private fun setupSwipeToDelete() {
         val itemTouchHelperCallback = object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
@@ -130,20 +243,24 @@ class CounterViewFragment : Fragment() {
 
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
                 val position = viewHolder.adapterPosition
-                val counterToDelete = counterList[position]
 
-                // Remove from Firestore
-                counterToDelete.counterId?.let {
-                    firestore.collection("counters").document(it)
-                        .delete()
-                        .addOnSuccessListener {
-                            toaster.showToast(getString(R.string.counter_delete_successful))
-                            counterList.removeAt(position)
-                            counterAdapter.notifyItemRemoved(position)
-                        }
-                        .addOnFailureListener { e ->
-                            toaster.showToast(getString(R.string.counter_delete_failed))
-                        }
+                if (position != RecyclerView.NO_POSITION) {
+                    val counterToDelete = counterList[position]
+
+                    //remove from Firestore
+                    counterToDelete.counterId?.let {
+                        firestore.collection("counters").document(it)
+                            .delete()
+                            .addOnSuccessListener {
+                                //remove from the list only after Firestore delete is successful
+                                //notify the adapter that the data has changed
+                                counterAdapter.notifyItemRemoved(position)
+                                toaster.showToast(getString(R.string.counter_delete_successful))
+                            }
+                            .addOnFailureListener { e ->
+                                toaster.showToast(getString(R.string.counter_delete_failed))
+                            }
+                    }
                 }
             }
